requirements.txt
numpy==2.2.5
pandas==2.2.3
PyYAML==6.0.2
scikit-learn==1.6.1
scipy==1.15.3